Move `DisplayFolder` to `:legacy:folder` module
