+++ legacy/folder/src/main/kotlin/app/k9mail/legacy/folder/DisplayFolder.kt
@@ -1,0 +1,8 @@
+package app.k9mail.legacy.folder
+
+data class DisplayFolder(
+    val folder: Folder,
+    val isInTopGroup: Boolean,
+    val unreadMessageCount: Int,
+    val starredMessageCount: Int,
+)
