+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/managefolders/ManageFoldersViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
 import app.k9mail.legacy.account.Account
-import com.fsck.k9.mailstore.DisplayFolder
+import app.k9mail.legacy.folder.DisplayFolder
 import com.fsck.k9.mailstore.FolderRepository
 
 class ManageFoldersViewModel(private val folderRepository: FolderRepository) : ViewModel() {
